Updated to Unity 4.6.6 and merged to the updated branch. The Rigged facet Character is in correctly, still with a hand animation issue which is being fixed at the moment.
I imported some hard surface shaders from the Unity webstore. Included the Facet character as a player character so you can get to him when the players switch through.

I also decreased the scale of the characters proportionally instead of the environment so that there was less work involved in getting the  environment to be more expansive, otherwise it would take way more steps.


Former-commit-id: e48784174ae55488260459e97d285ecebce9720e